OneStage/yolo/convert2Yolo/requirements.txt
Pillow==5.0.0
cycler==0.10.0
kiwisolver==1.0.1
matplotlib==2.2.2
numpy==1.14.3
Pillow==5.0.0
pyparsing==2.2.0
python-dateutil==2.7.2
pytz==2018.4
six==1.11.0
